+++ Products/SimpleGroupsManagement/Extensions/install.py
@@ -1,6 +1,9 @@
 # -*- coding: utf-8 -*-
 
+
 def uninstall(portal, reinstall=False):
     setup_tool = portal.portal_setup
     if not reinstall:
-        setup_tool.runAllImportStepsFromProfile('profile-Products.SimpleGroupsManagement:uninstall')
+        setup_tool.runAllImportStepsFromProfile(
+            'profile-Products.SimpleGroupsManagement:uninstall'
+        )