+++ Products/SimpleGroupsManagement/locales/rebuildAll.sh
@@ -4,4 +4,7 @@ DOMAIN="Products.SimpleGroupsManagement"
 
 i18ndude rebuild-pot --pot ${DOMAIN}.pot --create ${DOMAIN} ..
 i18ndude merge --pot ${DOMAIN}.pot --merge ${DOMAIN}-manual.pot
-i18ndude sync --pot ${DOMAIN}.pot ${DOMAIN}-??.po
+i18ndude sync --pot ${DOMAIN}.pot ./*/LC_MESSAGES/${DOMAIN}.po
+
+DOMAIN="plone"
+i18ndude sync --pot ${DOMAIN}.pot ./*/LC_MESSAGES/${DOMAIN}.po